news-market-agent/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
matplotlib==3.7.2
nltk==3.8.1
praw==7.7.1
schedule==1.2.0
python-dotenv==1.0.0
aiohttp==3.8.5
